requirements.txt
# Python dependencies
flask>=2.3.0
ultralytics>=8.3.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0